whisper/requirements.txt
torch >= 2.4.1, < 3.0.0
transformers <= 4.44.2, < 5.0.0
optimum-intel >= 1.19.0, < 2.0.0
openvino >=  2024.4.0, < 2025.0.0
# accelerate >= 0.34.2, < 1.0.0
# librosa >= 0.10.2.post1, < 1.0.0
# soundfile >= 0.12.1, < 1.0.0
# flash-attn >= 2.6.3, < 3.0.0

# TESTING
# datasets["audio"] >= 3.0.1, < 4.0.0